+++ zfit_physics/pdf.py
@@ -1,1 +1,4 @@
 from .models.pdf_argus import Argus
+from .models.pdf_relbw import RelativisticBreitWigner
+
+__all__ = ["Argus", "RelativisticBreitWigner"]